requirements.txt
appdirs==1.4.4
attrs==21.2.0
cattrs==1.7.1
certifi==2020.12.5
cfgv==3.3.0
chardet==4.0.0
distlib==0.3.2
Faker==8.5.0
filelock==3.0.12
identify==2.2.6
idna==2.10
iniconfig==1.1.1
nodeenv==1.6.0
packaging==20.9
pluggy==0.13.1
pre-commit==2.13.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.5
virtualenv==20.4.7
